pyproject.toml
[project]
name = "delivery-exception-system"
version = "0.1.0"
description = "AI-powered last mile delivery exception handling with LangGraph multi-agent system"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=0.3",
    "langchain-core>=0.3",
    "langchain-chroma>=0.2",
    "langchain-huggingface>=0.1",
    "langchain-text-splitters>=0.3",
    "langgraph>=0.3",
    "langsmith>=0.2",
    "pydantic>=2.0",
    "pypdf>=5.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
delivery-exception-system = "delivery_exception_system.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/delivery_exception_system"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
